move field ops to column file
